requirements.txt
python-dateutil
pytest
click
bson
werkzeug
flask==1.0
flask-sqlalchemy==2.1
pymongo
flask-jwt-extended
gunicorn
requests
wtforms==2.2.1
flask-wtf==0.14.2